requirements.txt
fastmcp>=3.0.0
aiosqlite>=0.19.0
PyYAML>=6.0
numpy>=1.24
sentence-transformers>=2.7.0
pypdf>=4.0.0
python-docx>=1.1.0
# Choose ONE:
# faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4
# Torch is required by sentence-transformers; install the correct build for your platform.
